successfully deferred quite a lot of interpolation calculations :)

+++ src/engine/camera.cpp
@@ -8,7 +8,7 @@ ThreeDL::Camera::Camera(const vec3& position, const vec3& rotation)
     recalculateVectors();
 }
 
-ThreeDL::Camera::Camera(const ThreeDL::vec3& position, const ThreeDL::vec3& rotation, const double fov)
+ThreeDL::Camera::Camera(const ThreeDL::vec3& position, const ThreeDL::vec3& rotation, const float fov)
     :   position_(position),
         rotation_(rotation),
         fov_(fov)
@@ -27,11 +27,11 @@ void ThreeDL::Camera::recalculateVectors() {
     right_.normalize();
 }
 
-void ThreeDL::Camera::moveForward(const double delta) {
+void ThreeDL::Camera::moveForward(const float delta) {
     position_ += forward_ * delta;
 }
 
-void ThreeDL::Camera::moveRight(const double delta) {
+void ThreeDL::Camera::moveRight(const float delta) {
     position_ += right_ * delta;
 }
 
@@ -41,17 +41,17 @@ void ThreeDL::Camera::rotate(const ThreeDL::vec3& delta) {
     roll(delta.z);
 }
 
-void ThreeDL::Camera::pan(const double delta) {
+void ThreeDL::Camera::pan(const float delta) {
     rotation_.y += delta;
     rotation_.y = fmin(rotation_.y, max_rotation_.y);
 }
 
-void ThreeDL::Camera::tilt(const double delta) {
+void ThreeDL::Camera::tilt(const float delta) {
     rotation_.x += delta;
     rotation_.x = fmin(rotation_.x, max_rotation_.x);
 }
 
-void ThreeDL::Camera::roll(const double delta) {
+void ThreeDL::Camera::roll(const float delta) {
     rotation_.z += delta;
     rotation_.z = fmin(rotation_.z, max_rotation_.z);
 }
